+++ src/main/java/com/ortec/gta/database/repository/RoleRepository.java
@@ -10,4 +10,4 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface RoleRepository extends CrudRepository<Role, Integer> {}
+public interface RoleRepository extends CrudRepository<Role, Long> {}
