+++ src/main/java/com/ortec/gta/database/repository/WeekRepository.java
@@ -10,7 +10,7 @@ import java.util.Set;
  * @Author: romain.pillot
  * @Date: 24/08/2017
  */
-public interface WeekRepository extends CrudRepository<Week, Integer> {
+public interface WeekRepository extends CrudRepository<Week, Long> {
     @Query("select w from Week w where w.user.id = ?1 and w.weekNumber = ?2 and w.year = ?3")
-    Set<Week> findByUserId(int id, int weekNumber, int year);
+    Set<Week> findByUserId(long id, int weekNumber, int year);
 }
